requirements.txt
fastapi
uvicorn[standard]
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
bcrypt
python-dotenv
SQLAlchemy==2.0.25
alembic
redis
celery
itsdangerous
requests
websockets
python-socketio[asgi]
aiofiles
